+++ psef_test/migrations/migration_7f5b06f81aef/__init__.py
@@ -1,0 +1,29 @@
+from .. import Tester
+
+
+def get_auto_tests(db):
+    return db.engine.execute('SELECT * from "AutoTest" ORDER BY id').fetchall()
+
+
+class UpgradeTester(Tester):
+    def load_data(self):
+        auto_tests = get_auto_tests(self.db)
+        assert auto_tests
+        assert all('prefer_teacher_revision' not in at for at in auto_tests)
+
+    def check(self):
+        auto_tests = get_auto_tests(self.db)
+        assert auto_tests
+        assert all(at['prefer_teacher_revision'] is False for at in auto_tests)
+
+
+class DowngradeTester(Tester):
+    def load_data(self):
+        auto_tests = get_auto_tests(self.db)
+        assert auto_tests
+        assert all('prefer_teacher_revision' in at for at in auto_tests)
+
+    def check(self):
+        auto_tests = get_auto_tests(self.db)
+        assert auto_tests
+        assert all('prefer_teacher_revision' not in at for at in auto_tests)